refactor: move all exceptions to common modulo core header
