--- source/modulo_core/include/modulo_core/exceptions/MessageTranslationException.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "modulo_core/exceptions/CoreException.hpp"
-
-namespace modulo_core::exceptions {
-
-/**
- * @class MessageTranslationException
- * @brief An exception class to notify that the translation of a ROS message failed.
- */
-class MessageTranslationException : public CoreException {
-public:
-  explicit MessageTranslationException(const std::string& msg) : CoreException("MessageTranslationException", msg) {}
-};
-}// namespace modulo_core::exceptions
-
-
